Adds msg_index to bridge evidence
Adds the `msg_index` field to the `bridge_evidence` type and related input/enum types in the GraphQL schema.

This field is important for uniquely identifying bridge evidence entries, in addition to the hash.
It also updates the metadata for `bridge_evidence` table to include `msg_index` in select permissions.